track if measurement thread is active
This patch adds a variable thread_active which is 1 while the
measurement thread is active, and is 0 otherwise. This state is tracked
manually. This is required to avoid the attempt of joining a non-created
thread.

According to pthread_equal(3) it is not acceptable to directly compare
two pthread_t, as they should be considered opaque. Hence, external
tracking whether a thread is active is required. (For this reason, the
hacky solution of comparing to 0 is not used, also it would probably
work.)

This patch further adds error checking for the final thread join. On
error, errno is printed with a debug message.